[Major] moves all random methods in a seperate file, adds the modlanguagesize const
